Changed movie table schema and csv file creation, cleaned up code, normalization of movie table
