+++ database/seedPostgresqlMovie.js
@@ -1,4 +1,4 @@
-const { Pool, Client } = require('pg');
+const { Pool} = require('pg');
 const path = require('path');
 const pool = new Pool({
   user: 'postgres',
@@ -12,32 +12,27 @@ const pool = new Pool({
   const client = await pool.connect();
 
   try {
-    console.time('timing seed start actor');
+    console.time('timing seed');
     // Transaction BEGIN!
     await client.query('BEGIN');
-    console.log('creating actorinfo table!');
+    console.log('creating movieinfo table!');
     await client.query(`
-      CREATE TABLE IF NOT EXISTS ActorInfo(
-        id NUMERIC NOT NULL,
-        name VARCHAR(100),
-        title VARCHAR(100),
-        role VARCHAR(100),
-        photo VARCHAR(100),
-        bio VARCHAR(2000),
-        filmography VARCHAR []
+      CREATE TABLE IF NOT EXISTS MovieInfo(
+        movieId NUMERIC NOT NULL,
+        actorId NUMERIC NOT NULL
         );
     `);
 
     console.log('writing to database!');
-    const copyPath = path.join(__dirname, '../actorData.csv');
+    const copyPath = path.join(__dirname, '../movieData.csv');
     await client.query(`
-      COPY ActorInfo FROM '${copyPath}' WITH (FORMAT CSV, HEADER);
+      COPY MovieInfo FROM '${copyPath}' WITH (FORMAT CSV, HEADER);
     `);
 
     console.log('commiting!');
     await client.query('COMMIT');
     // Transaction END!
-    console.timeEnd('timing seed end actor');
+    console.timeEnd('timing seed');
   } catch (e) {
     await client.query('ROLLBACK');
     console.log('error!');
@@ -49,7 +44,3 @@ const pool = new Pool({
 })().catch(e => console.error(e.stack));
 
 
-// CREATE TABLE IF NOT EXISTS MovieInfo(
-//   movieId NUMERIC NOT NULL,
-//   actors ARRAY
-//   );
