+++ package.json
@@ -12,9 +12,9 @@
     "react-dev": "webpack -d --watch",
     "server-dev": "nodemon server/index.js",
     "seed-db": "mongoimport --db fandangit --collection actors --drop --file database/fandangit-cast-2001.json",
-    "newSeed": "node --max_old_space_size=8048 ./database/movieSeed.js",
-    "seed:postgres:actor": "node --max-old-space-size=4096 ./database/pgSchemaActorInfo.js",
-    "seed:postgres:movie": "node --max-old-space-size=4096 ./database/pgSchemaMovieInfo.js"
+    "newSeed": "node --max_old_space_size=8048 ./database/writeMovieData.js",
+    "seed:postgres:actor": "node --max-old-space-size=4096 ./database/seedPostgresqlActor.js",
+    "seed:postgres:movie": "node --max-old-space-size=4096 ./database/seedPostgresqlMovie.js"
   },
   "dependencies": {
     "cors": "^2.8.5",
